requirements.txt
numpy
Pillow
colorlog
webcolors